+++ flow/__init__.py
@@ -1,2 +1,2 @@
 APP_NAME = "FlowMobile"
-APP_VERSION = "7.6.7"
+APP_VERSION = "7.6.8"